requirements.txt
python-dotenv
langchain-openai
langchain-core
langgraph
clickhouse-connect
